+++ nodejs/awsx/apigateway/index.ts
@@ -12,13 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from "./cluster";
-export * from "./network";
-
-import * as apigateway from "./apigateway";
-import * as autoscaling from "./autoscaling";
-import * as ec2 from "./ec2";
-import * as ecs from "./ecs";
-import * as elasticloadbalancingv2 from "./elasticloadbalancingv2";
-
-export { apigateway, autoscaling, ec2, ecs, elasticloadbalancingv2 };
+export * from "./api";